综述: 增强连接追踪能力并修复Content-Length类型限制 - 为DefaultPipeline、AdaptiveReadCompletionHandler和HttpConnection2添加唯一标识符uid，支持完整的连接生命周期追踪和调试 - 修复HttpDecodeUtil、HttpReceiveResponse、HttpReceiveResponseDecoder和AggregatorResponseFuture中Content-Length类型限制，从int改为long，支持超过2GB的HTTP响应体处理 - 在AggregatorResponseFuture中增加Content-Length溢出检查，超过Integer.MAX_VALUE时抛出异常 - 在HttpConnection2的响应处理中增加异常捕获机制，发生异常时自动关闭pipeline输入流，防止连接泄漏 - 为HttpConnection2添加toString方法，输出连接和pipeline的uid信息，便于日志追踪 - 注释HttpConnection、HttpConnection2、HttpConnection2Pool、HttpConnectionPool中的大量debug和error级别日志，减少生产环境日志噪音 - 注释CorsEncoder、HttpRespEncoder、HttpRequestPartDecoder、HttpResponsePartDecoder、ResourceProcessor、AbstractResourceEncoder等编解码器中的trace级别日志 - 注释SSLDecoder和SSLEncoder中的调试日志输出 - 删除临时文件ReverseApp.jar和开发计划.md文档

+++ src/main/java/cc/jfire/jnet/common/util/HttpDecodeUtil.java
@@ -106,9 +106,9 @@ public class HttpDecodeUtil
         headers.entrySet().stream().filter(entry -> entry.getKey().equalsIgnoreCase("Content-Type")).map(Map.Entry::getValue).findFirst().ifPresent(contentTypeConsumer);
     }
 
-    public static void findContentLength(Map<String, String> headers, Consumer<Integer> contentLengthConsumer)
+    public static void findContentLength(Map<String, String> headers, Consumer<Long> contentLengthConsumer)
     {
-        headers.entrySet().stream().filter(entry -> entry.getKey().equalsIgnoreCase("Content-Length")).map(entry -> Integer.valueOf(entry.getValue())).findFirst().ifPresent(contentLengthConsumer);
+        headers.entrySet().stream().filter(entry -> entry.getKey().equalsIgnoreCase("Content-Length")).map(entry -> Long.valueOf(entry.getValue())).findFirst().ifPresent(contentLengthConsumer);
     }
 
     public static String pureUrl(String url)
